refactor(app): old pipette stuff
