+++ app/src/organisms/Desktop/Devices/ConfigurePipette/ConfigErrorBanner.tsx
