+++ app/src/organisms/Desktop/Devices/PipetteCard/PipetteSettingsSlideout.tsx
@@ -2,8 +2,8 @@ import { useTranslation } from 'react-i18next'
 import { Flex } from '@opentrons/components'
 import { useUpdatePipetteSettingsMutation } from '@opentrons/react-api-client'
 import { Slideout } from '/app/atoms/Slideout'
-import { ConfigFormSubmitButton } from '/app/organisms/ConfigurePipette/ConfigFormSubmitButton'
-import { ConfigurePipette } from '/app/organisms/ConfigurePipette'
+import { ConfigFormSubmitButton } from '../ConfigurePipette/ConfigFormSubmitButton'
+import { ConfigurePipette } from '../ConfigurePipette'
 
 import type { PipetteModelSpecs } from '@opentrons/shared-data'
 import type { PipetteSettingsFieldsMap } from '@opentrons/api-client'
